requirements.txt
Flask
flask-socketio
google-generativeai